refactor(reconcilers): isolate OAuth/TLS reconciler code behind distro build tag
The deployment and service reconcilers mixed ODH-specific OAuth proxy
injection, SAR ConfigMap management, and TLS serving-cert annotations
with upstream kserve logic. This made upstream diffs hard to track and
created friction when porting changes in either direction.

Extracts all OpenShift-specific reconciler code into `_odh.go` companion
files gated by `//go:build distro`, leaving shared files identical (or
nearly so) to upstream kserve/kserve. The polymorphism points are
`buildDeployments()` and `buildServiceList()` - each has a minimal
`_default.go` stub for the upstream path and a full `_odh.go`
implementation for the distro path.

Factory parameter structs drop the distro-specific `ClientSet` and
`ResourceType` fields, matching upstream exactly. The `resourceType`
parameter is removed from `NewRawKubeReconciler` - the distro path
derives the component kind from `InferenceGraphLabel` on componentMeta
instead. The dead `ResourceType` enum is removed from constants.

Factory tests are split by build tag so the upstream test stays clean and
the distro variant sets up the full fixture (ISVC + oauthProxy configmap)
needed by the ODH deployment path.

Fixes RHOAIENG-57498
