+++ pkg/apis/serving/v1beta1/configmap.go
@@ -130,16 +130,6 @@ type IngressConfig struct {
 }
 
 // +kubebuilder:object:generate=false
-type OauthConfig struct {
-	Image                  string `json:"image"`
-	CpuLimit               string `json:"cpuLimit"`
-	CpuRequest             string `json:"cpuRequest"`
-	MemoryLimit            string `json:"memoryLimit"`
-	MemoryRequest          string `json:"memoryRequest"`
-	UpstreamTimeoutSeconds string `json:"upstreamTimeoutSeconds,omitempty"`
-}
-
-// +kubebuilder:object:generate=false
 type DeployConfig struct {
 	DefaultDeploymentMode     string                     `json:"defaultDeploymentMode,omitempty"`
 	DeploymentRolloutStrategy *DeploymentRolloutStrategy `json:"deploymentRolloutStrategy,omitempty"`
