+++ pkg/controller/v1beta1/inferenceservice/reconcilers/raw/raw_kube_reconciler.go
@@ -59,7 +59,6 @@ func NewRawKubeReconciler(ctx context.Context,
 	client client.Client,
 	clientset kubernetes.Interface,
 	scheme *runtime.Scheme,
-	resourceType constants.ResourceType,
 	componentMeta metav1.ObjectMeta,
 	workerComponentMeta metav1.ObjectMeta,
 	componentExt *v1beta1.ComponentExtensionSpec,
@@ -185,9 +184,7 @@ func NewRawKubeReconciler(ctx context.Context,
 		deploymentMode,
 		reconcilers.WorkloadReconcilerParams{
 			Client:              client,
-			ClientSet:           clientset,
 			Scheme:              scheme,
-			ResourceType:        resourceType,
 			ComponentMeta:       componentMeta,
 			WorkerComponentMeta: workerComponentMeta,
 			ComponentExt:        componentExt,
@@ -205,7 +202,6 @@ func NewRawKubeReconciler(ctx context.Context,
 		reconcilers.ServiceReconcilerParams{
 			Client:           client,
 			Scheme:           scheme,
-			ResourceType:     resourceType,
 			ComponentMeta:    componentMeta,
 			ComponentExt:     componentExt,
 			PodSpec:          podSpec,
