+++ pkg/controller/v1beta1/inferenceservice/reconcilers/service/service_reconciler_default.go
@@ -1,0 +1,34 @@
+//go:build !distro
+
+/*
+Copyright 2021 The KServe Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package service
+
+import (
+	corev1 "k8s.io/api/core/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+
+	"github.com/kserve/kserve/pkg/apis/serving/v1beta1"
+)
+
+// buildServiceList is the upstream (non-distro) implementation: creates services
+// with no OCP serving-cert annotations or OAuth proxy port rewrites.
+func buildServiceList(componentMeta metav1.ObjectMeta, componentExt *v1beta1.ComponentExtensionSpec,
+	podSpec *corev1.PodSpec, multiNodeEnabled bool, serviceConfig *v1beta1.ServiceConfig,
+) []*corev1.Service {
+	return createService(componentMeta, componentExt, podSpec, multiNodeEnabled, serviceConfig)
+}
