+++ pkg/openapi/swagger.json
@@ -4606,41 +4606,6 @@
         }
       }
     },
-    "v1beta1.OauthConfig": {
-      "type": "object",
-      "required": [
-        "image",
-        "cpuLimit",
-        "cpuRequest",
-        "memoryLimit",
-        "memoryRequest"
-      ],
-      "properties": {
-        "cpuLimit": {
-          "type": "string",
-          "default": ""
-        },
-        "cpuRequest": {
-          "type": "string",
-          "default": ""
-        },
-        "image": {
-          "type": "string",
-          "default": ""
-        },
-        "memoryLimit": {
-          "type": "string",
-          "default": ""
-        },
-        "memoryRequest": {
-          "type": "string",
-          "default": ""
-        },
-        "upstreamTimeoutSeconds": {
-          "type": "string"
-        }
-      }
-    },
     "v1beta1.OtelCollectorConfig": {
       "type": "object",
       "properties": {
